framework_core/pyproject.toml
[tool.poetry]
name = "syntonize-corekit"
version = "0.5.0"
description = "This is a project that contains common functions for all developers."
authors = ["lucasberlang <lcalvo@syntonize.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.11"
azure-monitor-opentelemetry = "^1.6.4"
logging = "^0.4.9.6"
types-pytz = "^2024.2.0.20241221"
azure-identity = {version = "^1.19.0", optional = true}
azure-keyvault-secrets = {version = "^4.9.0", optional = true}
pytest = "^8.3.4"
azure-data-tables = {version ="^12.6.0", optional = true}
azure-search-documents = {version = "^11.5.2", optional = true}
pyyaml = "^6.0.2"

[tool.poetry.extras]
azure = [
    "azure-identity",
    "azure-keyvault-secrets",
    "azure-data-tables",
    "azure-search-documents"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
